Refactor: moved the cops from lib/ into config/ and simplified names
The importing is where the naming convention comes in. The project itself doesn't actually provide you with a .rubocop.yml, so as long as you inherit from base then it's sweet.

Also, lib infers library code, while this is linting config. I am matching what I saw in the standard gem which has it's configs in a config folder.