Style: disabled RSpec/MultipleExpectations as it can be restricting
Tests should test a single thing, but sometimes it's more efficient to test a single thing through a series of assertions than needlessly splitting it. It always can be split, but I don't think it necessarily always improves readability and it can result in much slow CI runs if taken religiously.