style(layout): reset hash alignment to default after discussion
The main rationale for this originally was I think it looks pretty, but in the end, it breaks diffs really easily (as adding longer values edits the whole hash) and JSON/YAML are key aligned, so it's clearly not a readability issue. So, let's just go with the default, where the keys are next to the values.